move zuluru sync class to lib dir
