+++ server/lib/__init__.py
@@ -1,3 +1,4 @@
 from .stats_calculator import StatsCalculator
 from .stats_serializer import build_stats_response
 from .teams_serializer import build_teams_response
+from .zuluru_sync import ZuluruSync
